Two variations of the aria-hidden bookmarklet
This version of the aria-hidden bookmarklet has two options. One will force aria-hidden elements to be displayed. This is good when you have elements with display:none or visibility:hidden and you want to find them. The other version will show only those elements that are on the page. This is great for finding elements that should not have aria-hidden="true"